+++ src/endpoint-microservice/graphql/graphql-schema-converter/constants.ts
@@ -1,1 +1,3 @@
 export const DEFAULT_FIRST = 100;
+export const START_NAME_PATTERN = /[_a-zA-Z]/;
+export const CONTAIN_NAME_PATTERN = /^\w+$/;
