+++ src/endpoint-microservice/graphql/graphql-schema-converter/types/dateTimeType.ts
@@ -1,0 +1,5 @@
+import { GraphQLScalarType } from 'graphql/type';
+
+export const DateTimeType = new GraphQLScalarType({
+  name: 'DataTime',
+});
