+++ src/endpoint-microservice/graphql/graphql-schema-converter/types/jsonType.ts
@@ -1,0 +1,5 @@
+import { GraphQLScalarType } from 'graphql/type';
+
+export const JsonType = new GraphQLScalarType({
+  name: 'JSON',
+});
