+++ src/endpoint-microservice/graphql/graphql-schema-converter/types/serviceType.ts
@@ -1,0 +1,8 @@
+import { GraphQLObjectType, GraphQLString } from 'graphql/type';
+
+export const ServiceType = new GraphQLObjectType({
+  name: '_Service',
+  fields: {
+    sdl: { type: GraphQLString },
+  },
+});
