+++ src/endpoint-microservice/graphql/graphql-schema-converter/types/sortDirection.ts
@@ -1,0 +1,4 @@
+export enum SortDirection {
+  ASC = 'asc',
+  DESC = 'desc',
+}
