+++ src/endpoint-microservice/graphql/graphql-schema-converter/utils/isEmptyObject.ts
@@ -1,0 +1,25 @@
+import { JsonSchema } from 'src/endpoint-microservice/shared/schema';
+
+export const isEmptyObject = (schema: JsonSchema): boolean => {
+  if ('$ref' in schema) {
+    return false;
+  }
+
+  if (schema.type === 'object') {
+    const countProperties = Object.keys(schema.properties).length;
+
+    if (countProperties === 0) {
+      return true;
+    } else {
+      return Object.values(schema.properties).every((property) =>
+        isEmptyObject(property),
+      );
+    }
+  }
+
+  if (schema.type === 'array') {
+    return isEmptyObject(schema.items);
+  }
+
+  return false;
+};
